(remove space and add a comment)
The current version of Mister/TRS-80 doesn't support JV1 disk images of more than 80 tracks.
If such an image is inserted, it gets badly scrambled.
More than 80 tracks JV1 disk images do work with trstools and xtrs, the linux emulator, but fail here.

The reason is that in such case, the fdc code does change the disk parameters to double-side, which is not in the JV1 specs

With this PR, it seems to work with Newdos-80 and to some extent with trsdos2.3